Switch to flex containers where possible.
Long text in pill-hdrs now looks less awful, at the expense of unevenly
stretching the header bars.
This only happens on a small range of screens from md: to lg:, though.
The fix is to re-arrange the DOM s/t we have the pill-hdrs and the pills
at the same hierarchical level, and we arrange them so each row is
max-content (or something).
However, that sounds like a headache and I don't want to do it right
now.